requirements.txt
fastapi
uvicorn
PyPDF2
python-docx
nltk
transformers
language-tool-python
torch
sentence-transformers